requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.2.0
python-dotenv==1.0.0
numpy==1.26.4
geopy==2.4.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
streamlit==1.33.0
plotly==5.20.0
